transition sow to grpc and scaffold reap
Transitioned sow to use GRPC instead of rest, and as part of that moved
the http rest code to reap. This was done instead of using something
like GRPC gateway for learning means, but also to make it easier to
control how requests get modified before they're sent to end services.